+++ src/js/control-bar/control-bar.js
@@ -37,7 +37,8 @@ class ControlBar extends Component {
    */
   createEl() {
     return super.createEl('div', {
-      className: 'vjs-control-bar'
+      className: 'vjs-control-bar',
+      dir: 'ltr'
     }, {
       'role': 'group' // The control bar is a group, so it can contain menuitems
     });