+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * @llun fixed menus from throwing when focused when empty ([view](https://github.com/videojs/video.js/pull/3218))
+* @mister-ben added dir=ltr to control bar and loading spinner ([view](https://github.com/videojs/video.js/pull/3221))
 
 --------------------